Make new, performant deck analyzer.
The repo now contains perk sheet spoilers.

The old direction of brute-forcing through rolling modifiers
took several hours to do anything meaningful.
While starting simple, this new way is capable of analyzing
one of the simplest perk sheets in the game.
The new direction is going to be based in combinatorics,
should cut down on the time it takes to calculate results.